Add Express - Added Express - Split tmi.js code in to include file - Other misc refactoring
