+++ .env.sample
@@ -3,8 +3,8 @@ NODE_ENV=local
 
 LOGIT_LEVEL=compressed
 
-OAUTH_TOKEN=
 BOT_USERNAME=
+OAUTH_TOKEN=
 CHANNEL_NAME=
 
 PORT=1337
