+++ includes/_options.js
@@ -1,3 +1,5 @@
+require('dotenv').config();
+
 module.exports = {
     tmi: {
         client: {
